break up LinkCrud to reduce method count
