+++ gedbrowserng/src/main/java/org/schoellerfamily/gedbrowser/api/crud/LinkCrud.java
@@ -5,28 +5,11 @@ import java.util.List;
 
 import org.schoellerfamily.gedbrowser.api.datamodel.ApiAttribute;
 import org.schoellerfamily.gedbrowser.api.datamodel.ApiObject;
-import org.schoellerfamily.gedbrowser.persistence.mongo.gedconvert.GedObjectToGedDocumentMongoConverter;
-import org.schoellerfamily.gedbrowser.persistence.mongo.loader.GedDocumentFileLoader;
-import org.schoellerfamily.gedbrowser.persistence.mongo.repository.RepositoryManagerMongo;
 
 /**
  * @author Dick Schoeller
  */
-public interface LinkCrud {
-    /**
-     * @return the associated file loader
-     */
-    GedDocumentFileLoader getLoader();
-
-    /**
-     * @return the associated document converter
-     */
-    GedObjectToGedDocumentMongoConverter getConverter();
-
-    /**
-     * @return the associated repository manager
-     */
-    RepositoryManagerMongo getRepositoryManager();
+public interface LinkCrud extends NewCruds {
 
     /**
      * Check if this is a link to the source we are unlinking.
@@ -36,62 +19,6 @@ public interface LinkCrud {
      * @return true if matches
      */
     boolean isTheLinkWeAreLookingFor(ApiAttribute attribute, String id);
-
-    /**
-     * @return a new family CRUD object
-     */
-    default FamilyCrud familyCrud() {
-        return new FamilyCrud(getLoader(), getConverter(),
-                getRepositoryManager());
-    }
-
-    /**
-     * @return a new head CRUD object
-     */
-    default HeadCrud headCrud() {
-        return new HeadCrud(getLoader(), getConverter(),
-                getRepositoryManager());
-    }
-
-    /**
-     * @return a new note CRUD object
-     */
-    default NoteCrud noteCrud() {
-        return new NoteCrud(getLoader(), getConverter(),
-                getRepositoryManager());
-    }
-
-    /**
-     * @return a new person CRUD object
-     */
-    default PersonCrud personCrud() {
-        return new PersonCrud(getLoader(), getConverter(),
-                getRepositoryManager());
-    }
-
-    /**
-     * @return a new source CRUD object
-     */
-    default SourceCrud sourceCrud() {
-        return new SourceCrud(getLoader(), getConverter(),
-                getRepositoryManager());
-    }
-
-    /**
-     * @return a new submission CRUD object
-     */
-    default SubmissionCrud submissionCrud() {
-        return new SubmissionCrud(getLoader(), getConverter(),
-                getRepositoryManager());
-    }
-
-    /**
-     * @return a new submitter CRUD object
-     */
-    default SubmitterCrud submitterCrud() {
-        return new SubmitterCrud(getLoader(), getConverter(),
-                getRepositoryManager());
-    }
 
     /**
      * @param db the db to unlink from
